+++ src/structs/quaternion.rs
@@ -7,7 +7,7 @@ use std::ops::{Add, Sub, Mul, Div, Neg, AddAssign, SubAssign, MulAssign, DivAssi
 use std::iter::{FromIterator, IntoIterator};
 use rand::{Rand, Rng};
 use num::{Zero, One};
-use structs::{Vector3, Point3, Rotation3, Matrix3};
+use structs::{Vector3, Point3, Rotation3, Matrix3, Unit};
 use traits::operations::{ApproxEq, Inverse, PartialOrder, PartialOrdering, Axpy};
 use traits::structure::{Cast, Indexable, Iterable, IterableMut, Dimension, Shape, BaseFloat, BaseNum,
                         Bounded, Repeat};
@@ -31,24 +31,21 @@ pub struct Quaternion<N> {
     pub k: N
 }
 
-impl<N> Quaternion<N> {
+impl<N: Copy> Quaternion<N> {
     /// The vector part `(i, j, k)` of this quaternion.
     #[inline]
     pub fn vector(&self) -> &Vector3<N> {
-        // FIXME: do this require a `repr(C)` ?
-        unsafe {
-            mem::transmute(&self.i)
-        }
+        unsafe { mem::transmute(&self.i) }
     }
 
     /// The scalar part `w` of this quaternion.
     #[inline]
-    pub fn scalar(&self) -> &N {
-        &self.w
+    pub fn scalar(&self) -> N {
+        self.w
     }
 }
 
-impl<N: Neg<Output = N> + Copy> Quaternion<N> {
+impl<N: BaseNum + Neg<Output = N>> Quaternion<N> {
     /// Compute the conjugate of this quaternion.
     #[inline]
     pub fn conjugate(&self) -> Quaternion<N> {
@@ -64,7 +61,54 @@ impl<N: Neg<Output = N> + Copy> Quaternion<N> {
     }
 }
 
-impl<N: BaseFloat + ApproxEq<N>> Inverse for Quaternion<N> {
+impl<N: BaseFloat> Quaternion<N> {
+    /// Creates a new quaternion from its scalar and vector parts.
+    pub fn from_parts(scalar: N, vector: Vector3<N>) -> Quaternion<N> {
+        Quaternion::new(scalar, vector.x, vector.y, vector.z)
+    }
+
+    /// Creates a new quaternion from its polar decomposition.
+    ///
+    /// Note that `axis` is assumed to be a unit vector.
+    pub fn from_polar_decomposition(scalar: N, theta: N, axis: Unit<Vector3<N>>) -> Quaternion<N> {
+        let rot = UnitQuaternion::from_axisangle(axis, theta * ::cast(2.0));
+
+        rot.unwrap() * scalar
+    }
+
+    /// The polar decomposition of this quaternion.
+    ///
+    /// Returns, from left to right: the quaternion norm, the half rotation angle, the rotation
+    /// axis. If the rotation angle is zero, the rotation axis is set to the `y` axis.
+    pub fn polar_decomposition(&self) -> (N, N, Unit<Vector3<N>>) {
+        let nn = ::norm_squared(self);
+
+        let default_axis = Unit::from_unit_value_unchecked(Vector3::y());
+
+        if ApproxEq::approx_eq(&nn, &::zero()) {
+            (::zero(), ::zero(), default_axis)
+        }
+        else {
+            let n   = nn.sqrt();
+            let nq  = *self / n;
+            let v   = *self.vector();
+            let vnn = ::norm_squared(&v);
+
+            if ApproxEq::approx_eq(&vnn, &::zero()) {
+                (n, ::zero(), default_axis)
+            }
+            else {
+                let angle = self.scalar().acos();
+                let vn    = n.sqrt();
+                let axis  = Unit::from_unit_value_unchecked(v / vn);
+
+                (n, angle, axis)
+            }
+        }
+    }
+}
+
+impl<N: BaseFloat> Inverse for Quaternion<N> {
     #[inline]
     fn inverse(&self) -> Option<Quaternion<N>> {
         let mut res = *self;
@@ -86,17 +130,16 @@ impl<N: BaseFloat + ApproxEq<N>> Inverse for Quaternion<N> {
         }
         else {
             self.conjugate_mut();
-            self.w = self.w / norm_squared;
-            self.i = self.i / norm_squared;
-            self.j = self.j / norm_squared;
-            self.k = self.k / norm_squared;
+            *self /= norm_squared;
 
             true
         }
     }
 }
 
-impl<N: BaseFloat> Norm<N> for Quaternion<N> {
+impl<N: BaseFloat> Norm for Quaternion<N> {
+    type NormType = N;
+
     #[inline]
     fn norm_squared(&self) -> N {
         self.w * self.w + self.i * self.i + self.j * self.j + self.k * self.k
@@ -105,19 +148,27 @@ impl<N: BaseFloat> Norm<N> for Quaternion<N> {
     #[inline]
     fn normalize(&self) -> Quaternion<N> {
         let n = self.norm();
-        Quaternion::new(self.w / n, self.i / n, self.j / n, self.k / n)
+        *self / n
     }
 
     #[inline]
     fn normalize_mut(&mut self) -> N {
-        let n = Norm::norm(self);
-
-        self.w = self.w / n;
-        self.i = self.i / n;
-        self.j = self.j / n;
-        self.k = self.k / n;
+        let n = ::norm(self);
+        *self /= n;
 
         n
+    }
+
+    #[inline]
+    fn try_normalize(&self, min_norm: N) -> Option<Quaternion<N>> {
+        let n = ::norm(self);
+
+        if n <= min_norm {
+            None
+        }
+        else {
+            Some(*self / n)
+        }
     }
 }
 
@@ -143,7 +194,7 @@ impl<N> MulAssign<Quaternion<N>> for Quaternion<N>
     }
 }
 
-impl<N: ApproxEq<N> + BaseFloat> Div<Quaternion<N>> for Quaternion<N> {
+impl<N: BaseFloat> Div<Quaternion<N>> for Quaternion<N> {
     type Output = Quaternion<N>;
 
     #[inline]
@@ -152,11 +203,52 @@ impl<N: ApproxEq<N> + BaseFloat> Div<Quaternion<N>> for Quaternion<N> {
     }
 }
 
-impl<N: ApproxEq<N> + BaseFloat> DivAssign<Quaternion<N>> for Quaternion<N> {
+impl<N: BaseFloat> DivAssign<Quaternion<N>> for Quaternion<N> {
     #[inline]
     fn div_assign(&mut self, right: Quaternion<N>) {
         *self *= right.inverse().expect("Unable to invert the denominator.")
     }
+}
+
+impl<N: BaseFloat> Quaternion<N> {
+    /// Compute the exponential of a quaternion.
+    #[inline]
+    pub fn exp(&self) -> Self {
+        let v = *self.vector();
+        let nn = v.norm_squared();
+
+        if nn.is_zero() {
+            ::one()
+        }
+        else {
+            let n  = nn.sqrt();
+            let nv = v / n * n.sin();
+            Quaternion::from_parts(n.cos(), nv) * self.scalar().exp()
+        }
+    }
+
+    /// Compute the natural logarithm of a quaternion.
+    #[inline]
+    pub fn ln(&self) -> Self {
+        let n = self.norm();
+        let v = self.vector();
+        let s = self.scalar();
+
+        Quaternion::from_parts(n.ln(), v.normalize() *  (s / n).acos())
+    }
+
+    /// Raise the quaternion to a given floating power.
+    #[inline]
+    pub fn powf(&self, n: N) -> Self {
+        (self.ln() * n).exp()
+    }
+}
+
+impl<T> One for Quaternion<T> where T: Copy + One + Zero + Sub<T, Output = T> + Add<T, Output = T> {
+    #[inline]
+  fn one() -> Self {
+    Quaternion::new(T::one(), T::zero(), T::zero(), T::zero())
+  }
 }
 
 impl<N: fmt::Display> fmt::Display for Quaternion<N> {
@@ -165,122 +257,105 @@ impl<N: fmt::Display> fmt::Display for Quaternion<N> {
     }
 }
 
-/// A unit quaternion that can represent a 3D rotation.
-#[repr(C)]
-#[derive(Eq, PartialEq, RustcEncodable, RustcDecodable, Clone, Hash, Debug, Copy)]
-pub struct UnitQuaternion<N> {
-    q: Quaternion<N>
-}
+/// A unit quaternions. May be used to represent a rotation.
+pub type UnitQuaternion<N> = Unit<Quaternion<N>>;
+
+// /// A unit quaternion that can represent a 3D rotation.
+// #[repr(C)]
+// #[derive(Eq, PartialEq, RustcEncodable, RustcDecodable, Clone, Hash, Debug, Copy)]
+// pub struct UnitQuaternion<N> {
+//     q: Quaternion<N>
+// }
 
 impl<N: BaseFloat> UnitQuaternion<N> {
-    /// Creates a new unit quaternion from the axis-angle representation of a rotation.
+    /// Creates a new quaternion from a unit vector (the rotation axis) and an angle
+    /// (the rotation angle).
     #[inline]
-    pub fn new(axisangle: Vector3<N>) -> UnitQuaternion<N> {
-        let sqang = Norm::norm_squared(&axisangle);
+    pub fn from_axisangle(axis: Unit<Vector3<N>>, angle: N) -> UnitQuaternion<N> {
+        let (sang, cang) = (angle / ::cast(2.0)).sin_cos();
 
-        if ::is_zero(&sqang) {
-            ::one()
-        }
-        else {
-            let ang    = sqang.sqrt();
-            let (s, c) = (ang / Cast::from(2.0)).sin_cos();
+        let q = Quaternion::from_parts(cang, axis.unwrap() * sang);
+        Unit::from_unit_value_unchecked(q)
+    }
 
-            let s_ang = s / ang;
-
-            unsafe {
-                UnitQuaternion::new_with_unit_quaternion(
-                    Quaternion::new(
-                        c,
-                        axisangle.x * s_ang,
-                        axisangle.y * s_ang,
-                        axisangle.z * s_ang)
-                )
-            }
-        }
+    /// Same as `::from_axisangle` with the axis multiplied with the angle.
+    #[inline]
+    pub fn from_scaled_axis(axis: Vector3<N>) -> UnitQuaternion<N> {
+        let two: N = ::cast(2.0);
+        let q = Quaternion::from_parts(::zero(), axis * two).exp();
+        UnitQuaternion::from_unit_value_unchecked(q)
     }
 
     /// Creates a new unit quaternion from a quaternion.
     ///
     /// The input quaternion will be normalized.
     #[inline]
-    pub fn new_with_quaternion(q: Quaternion<N>) -> UnitQuaternion<N> {
-        UnitQuaternion { q: q.normalize() }
+    pub fn from_quaternion(q: &Quaternion<N>) -> UnitQuaternion<N> {
+        Unit::new(&q)
     }
 
     /// Creates a new unit quaternion from Euler angles.
     ///
     /// The primitive rotations are applied in order: 1 roll − 2 pitch − 3 yaw.
     #[inline]
-    pub fn new_with_euler_angles(roll: N, pitch: N, yaw: N) -> UnitQuaternion<N> {
+    pub fn from_euler_angles(roll: N, pitch: N, yaw: N) -> UnitQuaternion<N> {
         let (sr, cr) = (roll  * ::cast(0.5)).sin_cos();
         let (sp, cp) = (pitch * ::cast(0.5)).sin_cos();
         let (sy, cy) = (yaw   * ::cast(0.5)).sin_cos();
 
-        unsafe {
-            UnitQuaternion::new_with_unit_quaternion(
-                Quaternion::new(
+        let q = Quaternion::new(
                     cr * cp * cy + sr * sp * sy,
                     sr * cp * cy - cr * sp * sy,
                     cr * sp * cy + sr * cp * sy,
-                    cr * cp * sy - sr * sp * cy)
-            )
-        }
+                    cr * cp * sy - sr * sp * cy);
+
+        Unit::from_unit_value_unchecked(q)
+    }
+
+    /// The rotation angle of this unit quaternion.
+    #[inline]
+    pub fn angle(&self) -> N {
+        self.as_ref().scalar().acos()
+    }
+
+    /// The rotation axis of this unit quaternion or `None` if the rotation is zero.
+    #[inline]
+    pub fn axis(&self) -> Option<Unit<Vector3<N>>> {
+        Unit::try_new(self.as_ref().vector(), ::zero())
     }
 
     /// Builds a rotation matrix from this quaternion.
     pub fn to_rotation_matrix(&self) -> Rotation3<N> {
-        let ww = self.q.w * self.q.w;
-        let ii = self.q.i * self.q.i;
-        let jj = self.q.j * self.q.j;
-        let kk = self.q.k * self.q.k;
-        let ij = self.q.i * self.q.j * ::cast(2.0);
-        let wk = self.q.w * self.q.k * ::cast(2.0);
-        let wj = self.q.w * self.q.j * ::cast(2.0);
-        let ik = self.q.i * self.q.k * ::cast(2.0);
-        let jk = self.q.j * self.q.k * ::cast(2.0);
-        let wi = self.q.w * self.q.i * ::cast(2.0);
+        let ww = self.as_ref().w * self.as_ref().w;
+        let ii = self.as_ref().i * self.as_ref().i;
+        let jj = self.as_ref().j * self.as_ref().j;
+        let kk = self.as_ref().k * self.as_ref().k;
+        let ij = self.as_ref().i * self.as_ref().j * ::cast(2.0);
+        let wk = self.as_ref().w * self.as_ref().k * ::cast(2.0);
+        let wj = self.as_ref().w * self.as_ref().j * ::cast(2.0);
+        let ik = self.as_ref().i * self.as_ref().k * ::cast(2.0);
+        let jk = self.as_ref().j * self.as_ref().k * ::cast(2.0);
+        let wi = self.as_ref().w * self.as_ref().i * ::cast(2.0);
 
-        unsafe {
-            Rotation3::new_with_matrix(
-                Matrix3::new(
-                    ww + ii - jj - kk, ij - wk,           wj + ik,
-                    wk + ij,           ww - ii + jj - kk, jk - wi,
-                    ik - wj,           wi + jk,           ww - ii - jj + kk
-                )
+        Rotation3::new_with_matrix_unchecked(
+            Matrix3::new(
+                ww + ii - jj - kk, ij - wk,           wj + ik,
+                wk + ij,           ww - ii + jj - kk, jk - wi,
+                ik - wj,           wi + jk,           ww - ii - jj + kk
             )
-        }
-    }
-}
-
-
-impl<N> UnitQuaternion<N> {
-    /// Creates a new unit quaternion from a quaternion.
-    ///
-    /// This is unsafe because the input quaternion will not be normalized.
-    #[inline]
-    pub unsafe fn new_with_unit_quaternion(q: Quaternion<N>) -> UnitQuaternion<N> {
-        UnitQuaternion {
-            q: q
-        }
-    }
-
-    /// The `Quaternion` representation of this unit quaternion.
-    #[inline]
-    pub fn quaternion(&self) -> &Quaternion<N> {
-        &self.q
+        )
     }
 }
 
 impl<N: BaseNum> One for UnitQuaternion<N> {
     #[inline]
     fn one() -> UnitQuaternion<N> {
-        unsafe {
-            UnitQuaternion::new_with_unit_quaternion(Quaternion::new(::one(), ::zero(), ::zero(), ::zero()))
-        }
+        let one = Quaternion::new(::one(), ::zero(), ::zero(), ::zero());
+        UnitQuaternion::from_unit_value_unchecked(one)
     }
 }
 
-impl<N: Copy + Neg<Output = N>> Inverse for UnitQuaternion<N> {
+impl<N: BaseNum + Neg<Output = N>> Inverse for UnitQuaternion<N> {
     #[inline]
     fn inverse(&self) -> Option<UnitQuaternion<N>> {
         let mut cpy = *self;
@@ -291,7 +366,7 @@ impl<N: Copy + Neg<Output = N>> Inverse for UnitQuaternion<N> {
 
     #[inline]
     fn inverse_mut(&mut self) -> bool {
-        self.q.conjugate_mut();
+        *self = Unit::from_unit_value_unchecked(self.as_ref().conjugate());
 
         true
     }
@@ -300,7 +375,7 @@ impl<N: Copy + Neg<Output = N>> Inverse for UnitQuaternion<N> {
 impl<N: Rand + BaseFloat> Rand for UnitQuaternion<N> {
     #[inline]
     fn rand<R: Rng>(rng: &mut R) -> UnitQuaternion<N> {
-        UnitQuaternion::new(rng.gen())
+        UnitQuaternion::new(&rng.gen())
     }
 }
 
@@ -317,28 +392,28 @@ impl<N: ApproxEq<N>> ApproxEq<N> for UnitQuaternion<N> {
 
     #[inline]
     fn approx_eq_eps(&self, other: &UnitQuaternion<N>, eps: &N) -> bool {
-        ApproxEq::approx_eq_eps(&self.q, &other.q, eps)
+        ApproxEq::approx_eq_eps(self.as_ref(), other.as_ref(), eps)
     }
 
     #[inline]
     fn approx_eq_ulps(&self, other: &UnitQuaternion<N>, ulps: u32) -> bool {
-        ApproxEq::approx_eq_ulps(&self.q, &other.q, ulps)
+        ApproxEq::approx_eq_ulps(self.as_ref(), other.as_ref(), ulps)
     }
 }
 
-impl<N: BaseFloat + ApproxEq<N>> Div<UnitQuaternion<N>> for UnitQuaternion<N> {
+impl<N: BaseFloat> Div<UnitQuaternion<N>> for UnitQuaternion<N> {
     type Output = UnitQuaternion<N>;
 
     #[inline]
     fn div(self, other: UnitQuaternion<N>) -> UnitQuaternion<N> {
-        UnitQuaternion { q: self.q / other.q }
+        Unit::from_unit_value_unchecked(self.unwrap() / other.unwrap())
     }
 }
 
-impl<N: BaseFloat + ApproxEq<N>> DivAssign<UnitQuaternion<N>> for UnitQuaternion<N> {
+impl<N: BaseFloat> DivAssign<UnitQuaternion<N>> for UnitQuaternion<N> {
     #[inline]
     fn div_assign(&mut self, other: UnitQuaternion<N>) {
-        self.q /= other.q
+        *self = Unit::from_unit_value_unchecked(*self.as_ref() / *other.as_ref())
     }
 }
 
@@ -347,14 +422,14 @@ impl<N: BaseNum> Mul<UnitQuaternion<N>> for UnitQuaternion<N> {
 
     #[inline]
     fn mul(self, right: UnitQuaternion<N>) -> UnitQuaternion<N> {
-        UnitQuaternion { q: self.q * right.q }
+        Unit::from_unit_value_unchecked(self.unwrap() * right.unwrap())
     }
 }
 
 impl<N: BaseNum> MulAssign<UnitQuaternion<N>> for UnitQuaternion<N> {
     #[inline]
     fn mul_assign(&mut self, right: UnitQuaternion<N>) {
-        self.q *= right.q
+        *self = Unit::from_unit_value_unchecked(*self.as_ref() * *right.as_ref())
     }
 }
 
@@ -364,12 +439,9 @@ impl<N: BaseNum> Mul<Vector3<N>> for UnitQuaternion<N> {
     #[inline]
     fn mul(self, right: Vector3<N>) -> Vector3<N> {
         let two: N = ::one::<N>() + ::one();
-        let mut t = ::cross(self.q.vector(), &right);
-        t.x = t.x * two;
-        t.y = t.y * two;
-        t.z = t.z * two;
+        let t = ::cross(self.as_ref().vector(), &right);
 
-        Vector3::new(t.x * self.q.w, t.y * self.q.w, t.z * self.q.w) + ::cross(self.q.vector(), &t) + right
+        t * (two * self.as_ref().w) + ::cross(self.as_ref().vector(), &t) + right
     }
 }
 
@@ -421,14 +493,11 @@ impl<N: BaseNum + Neg<Output = N>> MulAssign<UnitQuaternion<N>> for Point3<N> {
 impl<N: BaseFloat> Rotation<Vector3<N>> for UnitQuaternion<N> {
     #[inline]
     fn rotation(&self) -> Vector3<N> {
-        let mut v = *self.q.vector();
-        let ang = v.normalize_mut().atan2(self.q.w) * ::cast(2.0);
-
-        if ::is_zero(&ang) {
-            ::zero()
+        if let Some(v) = self.axis() {
+            v.unwrap() * self.angle()
         }
         else {
-            Vector3::new(v.x * ang, v.y * ang, v.z * ang)
+            ::zero()
         }
     }
 
@@ -444,7 +513,7 @@ impl<N: BaseFloat> Rotation<Vector3<N>> for UnitQuaternion<N> {
 
     #[inline]
     fn append_rotation(&self, amount: &Vector3<N>) -> UnitQuaternion<N> {
-        *self * UnitQuaternion::new(*amount)
+        *self * UnitQuaternion::from_scaled_axis(*amount)
     }
 
     #[inline]
@@ -454,12 +523,12 @@ impl<N: BaseFloat> Rotation<Vector3<N>> for UnitQuaternion<N> {
 
     #[inline]
     fn prepend_rotation(&self, amount: &Vector3<N>) -> UnitQuaternion<N> {
-        UnitQuaternion::new(*amount) * *self
+        UnitQuaternion::from_scaled_axis(*amount) * *self
     }
 
     #[inline]
     fn set_rotation(&mut self, v: Vector3<N>) {
-        *self = UnitQuaternion::new(v)
+        *self = UnitQuaternion::from_scaled_axis(v);
     }
 }
 
@@ -496,7 +565,7 @@ impl<N: BaseNum + Neg<Output = N>> Rotate<Point3<N>> for UnitQuaternion<N> {
     }
 }
 
-impl<N: BaseFloat + ApproxEq<N>> RotationTo for UnitQuaternion<N> {
+impl<N: BaseFloat> RotationTo for UnitQuaternion<N> {
     type AngleType = N;
     type DeltaRotationType = UnitQuaternion<N>;
 
@@ -504,7 +573,7 @@ impl<N: BaseFloat + ApproxEq<N>> RotationTo for UnitQuaternion<N> {
     fn angle_to(&self, other: &Self) -> N {
         let delta = self.rotation_to(other);
 
-        delta.q.vector().norm().atan2(delta.q.w) * ::cast(2.0)
+        delta.as_ref().w.acos() * ::cast(2.0)
     }
 
     #[inline]
@@ -539,7 +608,8 @@ impl<N: BaseNum + Neg<Output = N>> Transform<Point3<N>> for UnitQuaternion<N> {
 
 impl<N: fmt::Display> fmt::Display for UnitQuaternion<N> {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "Unit quaternion {} − ({}, {}, {})", self.q.w, self.q.i, self.q.j, self.q.k)
+        write!(f, "Unit quaternion {} − ({}, {}, {})",
+               self.as_ref().w, self.as_ref().i, self.as_ref().j, self.as_ref().k)
     }
 }
 
@@ -558,42 +628,44 @@ impl<N> Dimension for UnitQuaternion<N> {
 #[cfg(feature="arbitrary")]
 impl<N: Arbitrary + BaseFloat> Arbitrary for UnitQuaternion<N> {
     fn arbitrary<G: Gen>(g: &mut G) -> UnitQuaternion<N> {
-        UnitQuaternion::new(Arbitrary::arbitrary(g))
+        UnitQuaternion::new(&Arbitrary::arbitrary(g))
     }
 }
 
-impl Quaternion<f32> {
-  /// Compute the exponential of a quaternion.
-  pub fn exp(&self) -> Self {
-    let v = *self.vector();
-    let nn = v.norm_squared();
-    
-    if nn.is_zero() {
-      ::one()
-    } else {
-      let n = nn.sqrt();
-      let nv = v / n * n.sin();
-      Quaternion::new(n.cos(), nv.x, nv.y, nv.z)
+impl<N: BaseFloat> UnitQuaternion<N> {
+    /// Compute the exponential of a quaternion.
+    ///
+    /// Note that this function yields a `Quaternion<N>` because it looses the unit property.
+    pub fn exp(&self) -> Quaternion<N> {
+        self.as_ref().exp()
     }
-  }
 
-  /// Compute the natural logarithm (a.k.a ln()) of a quaternion.
-  ///
-  /// Becareful, this function yields a `Quaternion<f32>`, losing the unit property.
-  pub fn log(&self) -> Self {
-    (Quaternion { w: 0., .. *self }).normalize() * self.w.acos()
-  }
+    /// Compute the natural logarithm of a quaternion.
+    ///
+    /// Note that this function yields a `Quaternion<N>` because it looses the unit property.  The
+    /// vector part of the return value corresponds to the axis-angle representation (divided by
+    /// 2.0) of this unit quaternion.
+    pub fn ln(&self) -> Quaternion<N> {
+        if let Some(v) = self.axis() {
+            Quaternion::from_parts(::zero(), v.unwrap() * self.angle())
+        }
+        else {
+            ::zero()
+        }
+    }
 
-  /// Raise the quaternion to a given floating power.
-  pub fn powf(&self, n: f32) -> Self {
-    (self.log() * n).exp()
-  }
-}
-
-impl<T> One for Quaternion<T> where T: Copy + One + Zero + Sub<T, Output = T> + Add<T, Output = T> {
-  fn one() -> Self {
-    Quaternion::new(T::one(), T::zero(), T::zero(), T::zero())
-  }
+    /// Raise this unit quaternion to a given floating power.
+    ///
+    /// If this unit quaternion represents a rotation by `theta`, then the resulting quaternion
+    /// rotates by `n * theta`.
+    pub fn powf(&self, n: N) -> Self {
+        if let Some(v) = self.axis() {
+            UnitQuaternion::from_axisangle(v, self.angle() * n)
+        }
+        else {
+            ::one()
+        }
+    }
 }
 
 componentwise_zero!(Quaternion, w, i, j, k);